moved settings to be part of bootstrap space content creation + then use from there when creating a space
